+++ lib/rdf/serializations/nquads_encoder.ex
@@ -3,7 +3,7 @@ defmodule RDF.NQuads.Encoder do
 
   use RDF.Serialization.Encoder
 
-  alias RDF.{Dataset, Graph, Quad, Statement, Triple}
+  alias RDF.{Dataset, Graph, Statement}
 
   @impl RDF.Serialization.Encoder
   @callback encode(Graph.t() | Dataset.t(), keyword | map) :: {:ok, String.t()} | {:error, any}
@@ -17,18 +17,17 @@ defmodule RDF.NQuads.Encoder do
     {:ok, if(result == "", do: result, else: "#{result}\n")}
   end
 
-  @spec statement({Statement.subject(), Statement.predicate(), Statement.object(), nil}) ::
-          String.t()
+  @spec statement(Statement.t()) :: String.t()
+  def statement(statement)
+
   def statement({subject, predicate, object, nil}) do
     statement({subject, predicate, object})
   end
 
-  @spec statement(Quad.t()) :: String.t()
   def statement({subject, predicate, object, graph}) do
     "#{term(subject)} #{term(predicate)} #{term(object)} #{term(graph)} ."
   end
 
-  @spec statement(Triple.t()) :: String.t()
   def statement({subject, predicate, object}) do
     "#{term(subject)} #{term(predicate)} #{term(object)} ."
   end